requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
click==8.3.1
Django==6.0.1
django-ninja==1.5.3
dnspython==2.8.0
elastic-transport==9.2.1
elasticsearch==9.3.0
email-validator==2.3.0
h11==0.16.0
idna==3.11
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
